+++ internal/productidentity/identity.go
@@ -17,7 +17,6 @@ const (
 	ProductURL               = "https://github.com/hdosys/herdr-sandbox"
 	ProductGUID              = "f845ea2a-ff3f-465b-95fa-49b1baf63b17"
 	UninstallKeyName         = "{F845EA2A-FF3F-465B-95FA-49B1BAF63B17}"
-	LegacyUninstallKeyName   = "HerdrSandbox"
 	InstallerMarkerName      = ".herdr-sandbox-installation.json"
 	QuietUninstallHelperName = ".herdr-sandbox-uninstall.ps1"
 	Copyright                = "Copyright (c) 2026 hdosys"